refactor: move activity in to components
