+++ apps/extension/src/entries/side-panel/components/Activities/Activity.tsx
