+++ apps/extension/src/entries/side-panel/components/Activities/DailyActivities.tsx
