+++ apps/extension/src/entries/side-panel/components/Activities/index.tsx
